Added User Login/Logout Test

+++ src/main/java/com/healthcare/model/entity/User.java
@@ -125,7 +125,7 @@ public @Data class User extends Audit implements Serializable {
 	private String expertDoctorTel;
 	@Column(name = "medical_condition")
 	private String medicalCondition;
-	private Integer status;
+	private long status;
 	@Column(name = "vacation_note")
 	private String vacationNote;
 	@Column(name = "vacation_start")
@@ -136,7 +136,7 @@ public @Data class User extends Audit implements Serializable {
 	@JoinColumn(name = "agency_id")
 	private Agency agency;
 	@Column(name = "status_second")
-	private Integer statusSecond;
+	private long statusSecond;
 	@Column(name = "remember_token")
 	private String rememberToken;
 	@ManyToOne
